Fix Accuracy and max combo being broken

+++ Script/Singleton/SettingsOperator.cs
@@ -25,8 +25,8 @@ public partial class SettingsOperator : Node
     public static int Level { get; set; } = 0;
     public static int OldCombo { get; set; } = 0;
     public static int OCombo { get; set; } = 0;
-    public static int OldAccuracy { get; set; } = 0;
-    public static int OAccuracy { get; set; } = 0;
+    public static float OldAccuracy { get; set; } = 0;
+    public static float OAccuracy { get; set; } = 0;
     public static bool JustPlayedScore {get; set;}
 
     public static string homedir = OS.GetUserDataDir().Replace("\\", "/");
